WIP initial refactor for sections
