+++ src/components/Sections/Hero/styles.js
