--- src/components/Sections/SectionBanner/index.js
@@ -1,1 +1,0 @@
-export { default as SectionBanner } from "./SectionBanner";
