--- src/components/Sections/SectionSpotlight/index.js
@@ -1,1 +1,0 @@
-export { default as SectionSpotlight } from "./SectionSpotlight";
