--- src/components/Sections/SectionStatement/index.js
@@ -1,1 +1,0 @@
-export { default as SectionStatement } from "./SectionStatement";
