--- src/components/Sections/SectionWrapper/index.js
@@ -1,1 +1,0 @@
-export { default as SectionWrapper } from "./SectionWrapper";
